Separate Cards component
